fix(release): split npm bootstrap preflight under the source length limit
